requirements.txt
discord.py
asyncpg
asyncio
datetime
pytz
yt_dlp